refactor: [STEP1] PR 리뷰 반영
- host, path 분리
- Respoonse 모델 네이밍 변경 및 파일 분리
- ModelTests 테스트 코드 수정:
